LayoutTests/imported/w3c/web-platform-tests/css/css-scoping/keyframes-002-expected.txt


FAIL @keyframes from the document should apply in the shadow tree assert_equals: expected 1 but got 0
